Fix sending back NotDone from renderd immediately if request doesn't fit into a rendering queue
RenderPrio, Render and RenderLow queues each only have a size of 32 entries. If these are full,
then requests overflow into the dirty queue. If this happens, renderd should send the NotDone
back immediately so that mod_tile doesn't wait for potentially a long time.

This however only worked for Render requests, but not RenderPrio or RenderLow. This patch fixes this.